feat: move to model.lesson package and compute distance proximity
Move SuitabilityScoreCalculator from utils to model.lesson package.

Update calculateSuitabilityScore to accept tutorLocation parameter.

Implement computeDistanceProximity to calculate distance between tutor and lesson locations.
